% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes.R
\name{construct_probes}
\alias{construct_probes}
\title{Concatenate Probe Content}
\usage{
construct_probes(vectors, S = 1:2, pR = 1:2)
}
\arguments{
\item{vectors}{Matrix containing the vectors to store in memory}

\item{S}{Stimulus presented on trial i}

\item{pR}{Response made on trial i-1}
}
\description{
When a stimulus is presented, we assume that the participant attempts to retrieve
the correct response given a probe composed of the current stimulus and the
participant's own response from the previous trial.
To do so, we applied Minerva 2's mechanism for cued recall; that is, we used
S_i R_{i-1} as a prompt to recover R_i.
}
