% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{construct_vectors}
\alias{construct_vectors}
\title{Generate vectors}
\usage{
construct_vectors(
  dimensionality = 20,
  features = c("stimuli", "responses"),
  max_cor = 0.4
)
}
\description{
To simulate the SRT task, we started by constructing 12 vectors, 1 to stand for each of the six stimulus positions and for each of the six possible responses.
Each vector was of dimensionality 20 with values of +1 and -1 selected at random with $p(+1)=p(-1)=.5$.
}
