% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{make_similarities}
\alias{make_similarities}
\title{Calculate similarities}
\usage{
make_similarities(
  probes,
  traces,
  index_probes,
  index_traces,
  preexperimental_traces,
  n_features = 40
)
}
\description{
function for one paricipant
}
